Implemented and used \ucwords macro
Added a marco to capitalize words of at least a minimum length. Care has
been taken that the argument is not expanded so you can use it with
embedded macros, such as \ucwords[4]{Macro are postponed: giassa\'i}.
The default argument specifies the minimun length of the words to
capitalize and it defaults to 3.

Overriden \section to use \ucwords on its argument: being it an arguable
cosmetic improvement, there is no need to have it in the document.